Combine SBS `maximal' and `minimal' DTBook grammars
Since we introduced the new more permissive schema for complicated
books, we've been using it together with the old strict schema for
simple books. To make this more explicit, we'll be requiring a special
proprietary version number on all books from now on. A DTBook that uses
the maximal grammar should have the version `2005-3-sbs', a DTBook that
uses the minimal grammar should have the version `2005-3-sbs-minimal'.

dtbook-2005-3-sbs-minimal.rnc is a copy of the old minimalSchema.rnc,
with a few modifications:
- version changed to 2005-3-sbs-minimal
- fixed pagenum id attribute datatype (NCName -> ID)

With the "externalRef trick" as explained on [http://norman.walsh.name/2004/07/25/xslt20],
I combine `dtbook-2005-3-sbs' and `dtbook-2005-3-sbs-minimal' into
`dtbook-2005-3-sbs-all'. This makes it possible to validate a DTBook
against either schema by referencing a single file.

A possible issue with this is that Oxygen applies the Schematron rules
embedded in `dtbook-2005-3-sbs', regardless of whether it is a minimal
or maximal DTBook. Luckily, in our case this doesn't matter because the
Schematron rules are about elements/attributes that aren't even allowed
by the minimal schema.